+++ packages/bsky/src/api/app/bsky/unspecced/getSuggestedOnboardingUsers.ts
@@ -26,7 +26,7 @@ export default function (server: Server, ctx: AppContext) {
   )
   server.add(app.bsky.unspecced.getSuggestedOnboardingUsers, {
     auth: ctx.authVerifier.standardOptional,
-    handler: async ({ auth, params, req }) => {
+    handler: async ({ auth, params, req, signal }) => {
       const viewer = auth.credentials.iss
       const labelers = ctx.reqLabelers(req)
       const hydrateCtx = await ctx.hydrator.createContext({
@@ -44,6 +44,7 @@ export default function (server: Server, ctx: AppContext) {
           ...params,
           hydrateCtx,
           headers,
+          signal,
         },
         ctx,
       )
@@ -70,6 +71,7 @@ const skeleton = async (input: SkeletonFnInput<Context, Params>) => {
     },
     {
       headers: params.headers,
+      signal: params.signal,
     },
   )
 }
@@ -133,6 +135,7 @@ type Context = {
 type Params = app.bsky.unspecced.getSuggestedOnboardingUsers.$Params & {
   hydrateCtx: HydrateCtx & { viewer: string | null }
   headers: Record<string, string>
+  signal: AbortSignal
   category?: string
 }